make visual test fail to start if port occupied instead of killing existing process on port

+++ tests/scripts/visual-coverage.sh
@@ -1,16 +1,24 @@
 #!/bin/bash
 
-# start Storybook in the same shell
-. ./tests/scripts/start-storybook.sh
+#check for existing process on port:6006
+process=`lsof -i :6006`
+if [[ -z $process ]]
+then
+  echo "port:6006 available - starting storybook"
+  # start Storybook in the same shell
+  . ./tests/scripts/start-storybook.sh
 
-# generate coverage from storybook test-runner
-test-storybook --coverage
+  # generate coverage from storybook test-runner
+  test-storybook --coverage
 
-# kill the locally served Storybook
-npx kill-port 6006
+  # kill the locally served Storybook
+  npx kill-port 6006
 
-# generate lcov coverage for visual tests from story book
-nyc report --reporter=lcov -t coverage/storybook --report-dir coverage/visual
+ # generate lcov coverage for visual tests from story book
+  nyc report --reporter=lcov -t coverage/storybook --report-dir coverage/visual
 
-cp coverage/storybook/coverage-storybook.json coverage/visual/coverage-storybook.json
-rm -r coverage/storybook
+  cp coverage/storybook/coverage-storybook.json coverage/visual/coverage-storybook.json
+  rm -r coverage/storybook
+else
+  echo -e "port:6006 in use - please kill current process before starting storybook: \n $process"
+fi
